--- data/classic/index.html
@@ -1,52 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-
-<head>
-  <title>Lane Pit Control</title>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" type="text/css" href="../style.css">
-  <link rel="icon" type="image/png" href="favicon.png">
-</head>
-
-<body>
-  <div class="topnav">
-    <label class="hamburger-menu">
-      <input type="checkbox">
-    </label>
-    <aside class="sidebar">
-      <nav>
-        <a href="../index.html">Main</a>
-        <a href="../soundboard/index.html">Soundboard</a>
-      </nav>
-    </aside>
-    <div class="header">
-      DRW Pit Caller - Classic
-    </div>
-  </div>
-  <div id="lanes">
-    <div class="card" id="lane1">
-      <h2>Lane 1</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(0, this.value)">
-      <button class="button" onclick="startPit(0)">Pit</button>
-    </div>
-    <div class="card" id="lane2">
-      <h2>Lane 2</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(1, this.value)">
-      <button class="button" onclick="startPit(1)">Pit</button>
-    </div>
-    <div class="card" id="lane3">
-      <h2>Lane 3</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(2, this.value)">
-      <button class="button" onclick="startPit(2)">Pit</button>
-    </div>
-    <div class="card" id="lane4">
-      <h2>Lane 4</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(3, this.value)">
-      <button class="button" onclick="startPit(3)">Pit</button>
-    </div>
-  </div>
-  <script src="script.js"></script>
-</body>
-
-</html>
