+++ data/index.html
@@ -11,55 +11,38 @@
 
 <body>
   <div class="topnav">
-    <label class="hamburger-menu">
-      <input type="checkbox">
-    </label>
-    <aside class="sidebar">
-      <nav>
-        <a href="classic/index.html">Classic</a>
-        <a href="soundboard/index.html">Soundboard</a>
-      </nav>
-    </aside>
     <div class="header">
-      DRW Pit Caller
+      Admin Page - DroneGP
     </div>
   </div>
   <div id="lanes">
     <div class="card" id="lane1">
       <h2>Lane 1</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(0, this.value)">
-      <button class="button" onclick="startPit(0)">Pit</button>
+      <input type="text" class="pilotName" placeholder="Team Name" oninput="updateLaneName(0, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 1" oninput="updatePilotName1(0, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 2" oninput="updatePilotName2(0, this.value)">
+      <!-- <button class="button" onclick="startPit(0)">Pit</button> -->
     </div>
     <div class="card" id="lane2">
       <h2>Lane 2</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(1, this.value)">
-      <button class="button" onclick="startPit(1)">Pit</button>
+      <input type="text" class="pilotName" placeholder="Team Name" oninput="updateLaneName(1, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 1" oninput="updatePilotName1(1, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 2" oninput="updatePilotName2(1, this.value)">
+      <!-- <button class="button" onclick="startPit(1)">Pit</button> -->
     </div>
     <div class="card" id="lane3">
       <h2>Lane 3</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(2, this.value)">
-      <button class="button" onclick="startPit(2)">Pit</button>
+      <input type="text" class="pilotName" placeholder="Team Name" oninput="updateLaneName(2, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 1" oninput="updatePilotName1(2, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 2" oninput="updatePilotName2(2, this.value)">
+      <!-- <button class="button" onclick="startPit(2)">Pit</button> -->
     </div>
     <div class="card" id="lane4">
       <h2>Lane 4</h2>
-      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(3, this.value)">
-      <button class="button" onclick="startPit(3)">Pit</button>
-    </div>
-    <div class="card">
-      <h2>Pit Stats</h2>
-      <table>
-        <tr>
-          <td>
-            <div class="stats" id="stats1">Lane 1: 0</div>
-            <div class="stats" id="stats2">Lane 2: 0</div>
-            <div class="stats" id="stats3">Lane 3: 0</div>
-            <div class="stats" id="stats4">Lane 4: 0</div>
-          </td>
-          <td>
-            <button class="button StatsButton" onclick="resetStats()">Reset</button>
-          </td>
-        </tr>
-      </table>      
+      <input type="text" class="pilotName" placeholder="Team Name" oninput="updateLaneName(3, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 1" oninput="updatePilotName1(3, this.value)">
+      <input type="text" class="pilotName" placeholder="Pilot Name 2" oninput="updatePilotName2(3, this.value)">
+      <!-- <button class="button" onclick="startPit(3)">Pit</button> -->
     </div>
   </div>
   <script src="script.js"></script>
