+++ data/lane1.html
@@ -1,0 +1,23 @@
+<!DOCTYPE HTML>
+<html>
+
+<head>
+  <title>Lane 1</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" type="text/css" href="style.css">
+  <link rel="icon" type="image/png" href="favicon.png">
+</head>
+
+<body>
+  <div id="lanes">
+    <div class="card" id="lane1">
+      <h2>Lane 1</h2>
+        <p>Pilot: </p>
+        <button class="button" onclick="startPit(0)">Pilot Swap</button>
+    </div>
+  </div>
+  <script src="script.js"></script>
+</body>
+
+</html>
