+++ data/lane2.html
@@ -1,0 +1,23 @@
+<!DOCTYPE HTML>
+<html>
+
+<head>
+  <title>Lane Pit Control</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" type="text/css" href="style.css">
+  <link rel="icon" type="image/png" href="favicon.png">
+</head>
+
+<body>
+  <div id="lanes">
+    <div class="card" id="lane2">
+      <h2>Lane 2</h2>
+      <input type="text" class="pilotName" placeholder="Pilot Name" oninput="updateLaneName(1, this.value)">
+      <button class="button" onclick="startPit(1)">Pit</button>
+    </div>
+  </div>
+  <script src="script.js"></script>
+</body>
+
+</html>
