+++ data/lane3.html
@@ -1,0 +1,22 @@
+<!DOCTYPE HTML>
+<html>
+
+<head>
+  <title>Lane 3</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" type="text/css" href="style.css">
+  <link rel="icon" type="image/png" href="favicon.png">
+</head>
+
+<body>
+  <div id="lanes">
+    <div class="card" id="lane3">
+      <h2>Team: </h2>
+      <button class="button" onclick="startPit(2)">Pilot Swap</button>
+    </div>
+  </div>
+  <script src="script.js"></script>
+</body>
+
+</html>
