+++ data/lane4.html
@@ -1,0 +1,22 @@
+<!DOCTYPE HTML>
+<html>
+
+<head>
+  <title>Lane 4</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" type="text/css" href="style.css">
+  <link rel="icon" type="image/png" href="favicon.png">
+</head>
+
+<body>
+  <div id="lanes">
+    <div class="card" id="lane4">
+      <h2>team:</h2>
+      <button class="button" onclick="startPit(3)">Pilot Swap</button>
+    </div>
+  </div>
+  <script src="script.js"></script>
+</body>
+
+</html>
