--- data/soundboard/index.html
@@ -1,62 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-
-<head>
-  <title>Lane Pit Control</title>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" type="text/css" href="../style.css">
-  <link rel="icon" type="image/png" href="../favicon.png">
-</head>
-
-<body>
-  <div class="topnav">
-    <label class="hamburger-menu">
-      <input type="checkbox">
-    </label>
-    <aside class="sidebar">
-      <nav>
-        <a href="../index.html">Main</a>
-        <a href="../classic/index.html">Classic</a>
-      </nav>
-    </aside>
-    <div class="header">
-      DRW Pit Caller - Soundboard
-    </div>
-  </div>
-  <div id="lanes">
-    <div class="card" id="lane1">
-      <h2>Message 1</h2>
-      <input type="text" class="pilotName" id="message1" placeholder="Enter Text">
-      <button class="button" onclick="playCustomMessage()">Play</button>
-    </div>
-    <div class="card" id="lane2">
-      <h2>Message 2</h2>
-      <input type="text" class="pilotName" id="message2" placeholder="Enter Text">
-      <button class="button" onclick="playCustomMessage()">Play</button>
-    </div>
-  </div>
-
-  <audio src="sounds/AirStrike.wav" id="AirStrike"></audio>
-  <audio src="sounds/CowMoo.wav" id="Moo"></audio>
-  <audio src="sounds/cymbals.wav" id="Cymbals"></audio>
-  <audio src="sounds/dumdum.wav" id="Dumdum"></audio>
-  <audio src="sounds/Explosion.wav" id="Explosion"></audio>
-  <audio src="sounds/NukeAnthem.wav" id="Anthem"></audio>
-  <audio src="sounds/Puke.wav" id="Puke"></audio>
-  <audio src="sounds/rubber_ducky.wav" id="RubberDucky"></audio>
-  <audio src="sounds/SheepBaa.wav" id="Baa"></audio>
-  <audio src="sounds/trumpet.wav" id="Trumpet"></audio>
-  <audio src="sounds/warning_horn.wav" id="WarningHorn"></audio>
-  <audio src="sounds/whah_whah.wav" id="Waahwaah"></audio>
-  <audio src="sounds/woow.wav" id="Woow"></audio>
-  <audio src="sounds/yay.wav" id="Yaay"></audio>
-  
-  <div class="soundcard" id="buttons">
-    <h2>Extra Sounds</h2>
-  </div>
-
-  <script src="script.js"></script>
-</body>
-
-</html>
