+++ data/style.css
@@ -260,3 +260,39 @@ body {
     /* border: 2px; */
     width: 100%;
 }
+
+
+.input-container h2 {
+    display: flex;
+    align-items: center;
+    margin: 0;
+    padding: 10px;
+    font-size: 2rem;
+    color: black;
+  }
+  .input-container input {
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    padding: 5px;
+    width: 200px;
+  }
+  .input-container button {
+    padding: 10px 0px;
+    font-size: 24px;
+    text-align: center;
+    width: 120px;
+    outline: none;
+    color: black;
+    background-color: yellow;
+    border: 4px solid black;
+    border-radius: 5px;
+    transition-duration: 0.3s;
+    user-select: none;
+  }
+  .input-container editButton:hover {
+    color: yellow;
+  }
+  .input-container editButton svg {
+    width: 20%;
+    height: 20%;
+  }
